+++ routes/DatasetRouter.js
@@ -16,8 +16,8 @@ class DatasetRouter {
         this.router.post('/subscriptions', authorize, this.datasetController.getMySubscriptions.bind(this.datasetController))
         this.router.post('/viewdataset', authorize, this.datasetController.viewDataset.bind(this.datasetController))
         this.router.post('/findsimilardatasets', authorize, this.datasetController.findSimilarDatasets.bind(this.datasetController))
-        this.router.get('/datasetpreview/:datasetId', this.datasetController.datasetPreview.bind(this.datasetController))
-        this.router.get('/datasetfullview/:datasetId/:subscriptionId', this.datasetController.datasetFullview.bind(this.datasetController))
+        this.router.get('/metadataapi/:datasetId', this.datasetController.getMetadata.bind(this.datasetController))
+        this.router.get('/dataapi/:datasetId/:subscriptionId', this.datasetController.getData.bind(this.datasetController))
         this.router.post('/createdataset',
             [
                 check('name', 'Name must not be empty').notEmpty(),